requirements.txt
flask==2.3.3
flask-restx==1.2.0
flask-cors==4.0.0
langchain==0.0.350
langchain-openai==0.0.5
langchain-community==0.0.10
psycopg2-binary==2.9.8
faiss-cpu==1.7.4
python-dotenv==1.0.0
requests==2.31.0